features/vector-search/requirements.txt
# API Framework
fastapi>=0.104.0              # Modern API framework with auto-generated OpenAPI schemas
uvicorn[standard]>=0.24.0     # High-performance ASGI server
pydantic>=2.0.0               # Data validation

# Document Processing
python-docx>=1.0.0            # Word document extraction
openpyxl>=3.1.0               # Excel spreadsheet parsing
python-pptx>=0.6.23           # PowerPoint slide extraction
pypdf>=3.17.0                 # PDF text extraction
markdown>=3.5.0               # Markdown processing

# Vector Database & Embeddings
chromadb>=0.4.18              # Embedded vector database
openai>=1.12.0                # OpenAI embeddings API (latest stable)
httpx>=0.27.0                 # HTTP client (compatible with openai)
tiktoken>=0.5.0               # Token counting for smart chunking

# Utilities
python-dotenv>=1.0.0          # Environment variable management

